helper autocomplete selection function
