+++ libs/expo/betterangels/src/lib/helpers/index.ts
@@ -3,3 +3,4 @@ export { default as formatSSN } from './formatSSN';
 export { default as generatePublicNote } from './generatePublicNote';
 export { default as handleEmailPress } from './handleEmailPress';
 export { default as searchPlacesInCalifornia } from './searchPlacesInCalifornia';
+export { default as selectAutocompletePlace } from './selectAutocompletePlace';
